+++ src/main/kotlin/icu/windea/pls/lang/util/ParadoxSpriteManager.kt
@@ -3,13 +3,14 @@ package icu.windea.pls.lang.util
 import icu.windea.pls.ep.data.*
 import icu.windea.pls.lang.*
 import icu.windea.pls.lang.util.image.*
+import icu.windea.pls.model.*
 import icu.windea.pls.script.psi.*
 
 object ParadoxSpriteManager {
-    fun getFrameInfo(sprite: ParadoxScriptDefinitionElement): FrameInfo? {
+    fun getFrameInfo(sprite: ParadoxScriptDefinitionElement): ImageFrameInfo? {
         val data = sprite.getData<ParadoxSpriteData>() ?: return null
         val frame = if (data.spriteSheetSpriteType != null) data.defaultFrame else null
         val frames = data.noOfFrames
-        return FrameInfo.of(frame, frames)
+        return ImageFrameInfo.of(frame, frames)
     }
 }