+++ src/main/kotlin/icu/windea/pls/lang/util/image/ParadoxImageResolver.kt
@@ -6,6 +6,7 @@ import com.intellij.openapi.progress.*
 import com.intellij.openapi.project.*
 import com.intellij.openapi.vfs.*
 import com.intellij.psi.*
+import icu.windea.pls.*
 import icu.windea.pls.core.*
 import icu.windea.pls.dds.*
 import icu.windea.pls.ep.data.*
@@ -16,6 +17,8 @@ import icu.windea.pls.model.*
 import icu.windea.pls.script.psi.*
 import org.intellij.images.fileTypes.impl.*
 import java.lang.invoke.*
+import java.nio.file.*
+import java.util.concurrent.atomic.*
 import kotlin.io.path.*
 
 object ParadoxImageResolver {
@@ -27,7 +30,7 @@ object ParadoxImageResolver {
     /**
      * 基于定义解析图片的路径。接受类型不为`sprite`的定义。返回用于渲染的图片的绝对路径。
      */
-    fun resolveUrlByDefinition(definition: ParadoxScriptDefinitionElement, frameInfo: FrameInfo? = null): String? {
+    fun resolveUrlByDefinition(definition: ParadoxScriptDefinitionElement, frameInfo: ImageFrameInfo? = null): String? {
         val definitionInfo = definition.definitionInfo ?: return null
         val newFrameInfo = when {
             frameInfo == null -> null
@@ -53,7 +56,7 @@ object ParadoxImageResolver {
     /**
      * 基于文件解析图片的路径。返回用于渲染的图片的绝对路径。
      */
-    fun resolveUrlByFile(file: VirtualFile, frameInfo: FrameInfo? = null): String? {
+    fun resolveUrlByFile(file: VirtualFile, frameInfo: ImageFrameInfo? = null): String? {
         try {
             //如果无法解析为png文件地址，则返回默认的地址
             val url = doResolveUrlByFile(file, frameInfo)
@@ -70,7 +73,7 @@ object ParadoxImageResolver {
     /**
      * 基于文件路径解析图片的路径。输入的文件路径需要相对于游戏或模组的根目录。返回用于渲染的图片的绝对路径。
      */
-    fun resolveUrlByFilePath(filePath: String, project: Project, frameInfo: FrameInfo? = null): String? {
+    fun resolveUrlByFilePath(filePath: String, project: Project, frameInfo: ImageFrameInfo? = null): String? {
         try {
             //如果无法解析为png文件地址，则返回默认的地址
             val url = doResolveUrlByFilePath(filePath, project, frameInfo)
@@ -84,7 +87,7 @@ object ParadoxImageResolver {
         }
     }
 
-    private fun doResolveUrlByDefinition(definition: ParadoxScriptDefinitionElement, definitionInfo: ParadoxDefinitionInfo, frameInfo: FrameInfo?): String? {
+    private fun doResolveUrlByDefinition(definition: ParadoxScriptDefinitionElement, definitionInfo: ParadoxDefinitionInfo, frameInfo: ImageFrameInfo?): String? {
         //兼容definition不是sprite的情况
         val resolved = runReadAction {
             definitionInfo.primaryImages.firstNotNullOfOrNull {
@@ -95,7 +98,7 @@ object ParadoxImageResolver {
         return doResolveUrlByFile(resolvedFile.virtualFile, resolved.frameInfo)
     }
 
-    private fun doResolveUrlByFile(file: VirtualFile, frameInfo: FrameInfo?): String? {
+    private fun doResolveUrlByFile(file: VirtualFile, frameInfo: ImageFrameInfo?): String? {
         val fileType = file.fileType
         if (fileType == ImageFileType.INSTANCE && file.extension?.lowercase().let { it == "png" }) {
             //accept png file
@@ -109,18 +112,30 @@ object ParadoxImageResolver {
         return ParadoxDdsResolver.resolveUrl(ddsAbsPath, ddsRelPath, frameInfo)
     }
 
-    private fun doResolveUrlByFilePath(filePath: String, project: Project, frameInfo: FrameInfo?): String? {
+    private fun doResolveUrlByFilePath(filePath: String, project: Project, frameInfo: ImageFrameInfo?): String? {
         val file = ParadoxFilePathSearch.search(filePath, null, selector(project).file()).find() ?: return null
         return doResolveUrlByFile(file, frameInfo)
     }
 
-    fun getDefaultUrl(): String {
-        return ParadoxDdsResolver.getUnknownPngUrl()
-    }
-
-    fun getPngFile(ddsFile: VirtualFile, frameInfo: FrameInfo? = null): VirtualFile? {
+    fun getPngFile(ddsFile: VirtualFile, frameInfo: ImageFrameInfo? = null): VirtualFile? {
         if (ddsFile.fileType != DdsFileType) return null // input file must be a dds file
         val absPngPath = doResolveUrlByFile(ddsFile, frameInfo) ?: return null
         return VfsUtil.findFile(absPngPath.toPath(), true)
+    }
+
+    fun getDefaultUrl(): String {
+        return getUnknownPngUrl()
+    }
+
+    private val clearUnknownPng = AtomicBoolean(true)
+
+    fun getUnknownPngUrl(): String {
+        //如果首次获取或者图片路径不存在，则将jar包中的unknown.png复制到~/.pls/images中
+        if (clearUnknownPng.getAndSet(false) || PlsConstants.Paths.unknownPngPath.notExists()) {
+            PlsConstants.Paths.unknownPngClasspathUrl.openStream().use { inputStream ->
+                Files.copy(inputStream, PlsConstants.Paths.unknownPngPath, StandardCopyOption.REPLACE_EXISTING)
+            }
+        }
+        return PlsConstants.Paths.unknownPngPath.toString()
     }
 }